Remove build dependency "pytest-runner" from setup.py
The build dependency "pytest-runner" (specified with "setup_requires"
in setup.py) causes pytest-runner to be installed on every invocation of
"python setup.py <command>". This no longer works on Python versions
older than 3.6 because pytest-runner has a dependency to the latest
version of setuptools_scm which recently dropped support for Python
older than 3.6. Now the CI of many Python projects is broken because of
this incompatibility.

Since usually we don't run tests with "python setup.py test", we
actually don't need this dependency anyway. So it should be safe to
simply remove the dependency. Running tests with the "pytest" command
will still work as usual.